1001.1: Separate match summary and battle components properly
